Post DataUpdateEvent on database modified
On deleting, inserting or updating a row in the database,
DataUpdateEvent is posted so that subscribers can reload the data.

Made some changes in creation of CouponTable, ie, defaults are no
longer specified in the create table statement. App ensures that all
values for coupon object is present and valid.