Fix add new google account bug
On adding a new account to the device and then trying to import or
export coupons, the sync task always failed. To prevent this, a flag is
used to keep track of any failed connection attempts. If true, this
implies that a new account was added to the device. In that case, when
the connection is established, it is disconnected and a request to
reconnect is made to successfully perform sync tasks.