Fix CouponFragment issue in tablets
In tablets, on viewing a coupon selected from the widget, each
orientation change led to recreation of a new instance of that fragment
added on the top of the fragment backstack leading to multiple copies.

Fixed this by using a boolean to track loading of coupon fragments from
widgets. If it is loaded this is set to false and prevents reloading of
the same fragment on orientation change.